% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFLP_plot_Evanno.R
\name{AFLP_plot_Evanno}
\alias{AFLP_plot_Evanno}
\title{Evanno plot function}
\usage{
AFLP_plot_Evanno(structure_output_path)
}
\arguments{
\item{structure_output_path}{path for the directory with structure output files}
}
\description{
Plot Evanno delta K plot from AFLP data, ggplot-based so can be styled with additional ggplot arguments, eg. AFLP_plot_Evanno + theme(axis.text = element_text(size = 10))
}
\examples{
structure_output_path <- "struct-new/output/Pmi"
AFLP_plot_Evanno(structure_output_path)
}
