% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFLP_plot_NJtree.R
\name{AFLP_plot_NJtree}
\alias{AFLP_plot_NJtree}
\title{NJ tree plot function}
\usage{
AFLP_plot_NJtree(
  structure_input_path,
  population_data_path = NULL,
  strata = 4,
  remove = NULL,
  labels = TRUE
)
}
\arguments{
\item{structure_input_path}{AFLP matrix where rows reperesent individuals and columns the markers, no header, first column are individual names in the format Population-Individual (eg. POP1-1), if any additional column should be ommited (eg. with population number) it should be passed using the 'remove' argument. Assumes that the first row should be removed and duplicated rows as well (that's how the structure input is usually coded for AFLP).}

\item{population_data_path}{path for the population data file (optional) where the first column is the population, second is  and the second the strata. Can have a header as long it does not contain any of the population names!}

\item{strata}{coulumn in the population data file used for data stratification}

\item{remove}{vector of colums to be removed, defalults to usual population column in structure files. Set as remove=NULL if no column should be removed.}

\item{labels}{should tip labels be printed or not}
}
\description{
Plot Jaccard distance based NJ tree from AFLP data (individuals with missing data are removed!), ggplot-based so can be styled with additional ggplot arguments, eg. AFLP_plot_NJtree + theme(axis.text = element_text(size = 10))
}
\examples{
structure_input_path <- "test_data/structure/Aal_carp_structure-inputD.txt"
population_data_path <- "test_data/structure/popdata.txt"
AFLP_plot_NJtree(structure_input_path, population_data_path, strata=5, remove=NULL)
}
