% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Evanno.R
\name{plot_Evanno}
\alias{plot_Evanno}
\title{Evanno plot function}
\usage{
plot_Evanno(structure_output_path, full = FALSE)
}
\arguments{
\item{structure_output_path}{path for the directory with structure output files}

\item{full}{should also other plots be generated: mean likelihood L(K), its rate of change L'(K)=L(K)), and absolute values of the second order rate of change |L''(K)|=|L'(K+1)-L'(K)| ?}
}
\description{
Plots Evanno delta K plot from structure output data, ggplot-based so can be styled with additional ggplot arguments, eg. plot_Evanno + theme(axis.text = element_text(size = 10))
}
\details{
Plots are based on Evanno (Evanno, G., Regnaut, S., Goudet, J. (2005). Detecting the number of clusters of individuals using the software STRUCTURE: a simulation study. Molecular ecology, 14(8), 2611-2620) method, data is generated by the pophelper package (Francis, R. M. (2017). POPHELPER: an R package and web app to analyse and visualize population structure. Molecular Ecology Resources, 17(1), 27-32. DOI: 10.1111/1755-0998.12509).
}
\examples{
structure_output_path <- "test_data/structure/Aal_admix_out"
plot_Evanno(structure_output_path)
}
