% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_str_mapdata.R
\name{prepare_str_mapdata}
\alias{prepare_str_mapdata}
\title{Prepare AFLP data for mapmixture}
\usage{
prepare_str_mapdata(structure_output_path, population_data_path)
}
\arguments{
\item{structure_output_path}{path for the directory with structure output files}

\item{population_data_path}{path for the population data file where the first column is the population and the second the strata. Can have a header as long it does not contain any of the population names!}

\item{choices}{vector containing the PCoA axes to be plotted}

\item{remove}{vector of colums to be removed}
}
\description{
Prepare data for plotting georgraphical map with structure piecharts from AFLP data using the mapmixture package, ggplot-based
}
\examples{
population_data_path <- "population_data.txt"
structure_output_path <- "struct-new/output/Pmi"
strmapdata <- prepare_str_mapdata(structure_output_path, population_data_path)
boundary <- c(xmin=5, xmax=27.5, ymin=43.5, ymax=50)
earth <- terra::rast("~/GIS/NE1_HR_LC_SR_W_DR/NE1_HR_LC_SR_W_DR.tif")
gg_color_hue <- function(n) {
  hues = seq(15, 375, length = n + 1)
  hcl(h = hues, l = 65, c = 100)[1:n]
}
k <- 2
mapmixture(strmapdata$str[[k]], strmapdata$pop, pie_size = 0.5, basemap = earth, boundary = boundary, cluster_cols = gg_color_hue(k), arrow_position = "tr")
}
