% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFLP_prepare_eems.R
\name{AFLP_prepare_eems}
\alias{AFLP_prepare_eems}
\title{Prepare EEMS input}
\usage{
AFLP_prepare_eems(species, aflp_matrix_path, population_data_path, remove = 2)
}
\arguments{
\item{species}{name of the output files}

\item{aflp_matrix_path}{AFLP matrix where rows reperesent individuals and columns the markers, no header, first column are individual names in the format Population-Individual (eg. POP1-1), if any additional column should be ommited (eg. with population number) it should be passed using the 'remove' argument.}

\item{population_data_path}{path for the population data file where the first column is the population, second the strata, third y and fourth z coordinate. Can have a header as long it does not contain any of the population names!}

\item{remove}{vector of colums to be removed}
}
\description{
Prepare input files for eems. Distances are calculated using jaccard distance, individuals with the same distances are removed (closely related or clonal individuals)
}
\examples{

aflp_matrix_path <- "struct-new/original_input/Cal_Cal_Structure.txt"
population_data_path <- "population_data.txt"
AFLP_plot_structure_map(aflp_matrix_path, population_data_path, c(2,3))AFLP_plot_PCoA(aflp_matrix_path=, population_data_path=NULL, choices=c(1,2))
}
