% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_str_mapdata.R
\name{prepare_str_mapdata}
\alias{prepare_str_mapdata}
\title{Prepare AFLP data for mapmixture}
\usage{
prepare_str_mapdata(
  structure_output_path,
  structure_input_path,
  population_data_path,
  filetype = "structure"
)
}
\arguments{
\item{structure_output_path}{path for the directory with structure output files}

\item{structure_input_path}{path for the structure input file or plink .fam file}

\item{population_data_path}{path for the population data file where the first column is the population. Can have a header as long it does not contain any of the population names!}

\item{filetype}{type of the input data ('auto', 'structure','tess2','baps','basic' or 'clumpp')}

\item{remove}{vector of colums to be removed}
}
\description{
Prepare data for plotting georgraphical map with structure piecharts from AFLP data using the mapmixture package, ggplot-based
}
\examples{
# For faststructure data:
population_data_path <- "test_data/fastStructure/Pta_population_data.txt"
structure_output_path <- "test_data/fastStructure/meanQ"
structure_input_path <- "test_data/fastStructure/Pta_plink.fam"
strmapdata <- prepare_str_mapdata(structure_output_path, structure_input_path, population_data_path, filetype="basic")
boundary <- c(xmin=18, xmax=21, ymin=48.5, ymax=49.5)
earth <- terra::rast("~/GIS/NE1_HR_LC_SR_W_DR/NE1_HR_LC_SR_W_DR.tif")
gg_color_hue <- function(n) {
  hues = seq(15, 375, length = n + 1)
  hcl(h = hues, l = 65, c = 100)[1:n]
}
k <- 3
mapmixture(strmapdata$str[[as.character(k)]], strmapdata$pop, pie_size = 0.1, basemap = earth, boundary = boundary, cluster_cols = gg_color_hue(k), arrow_position = "tr")

# For structure data:
population_data_path <- "test_data/structure/popdata.txt"
structure_output_path <- "test_data/structure/Aal_admix_out"
structure_input_path <- "test_data/structure/Aal_carp_structure-inputD.txt"
strmapdata <- prepare_str_mapdata(structure_output_path, structure_input_path, population_data_path, filetype="structure")
boundary <- c(xmin=16, xmax=27.5, ymin=43.5, ymax=50)
earth <- terra::rast("~/GIS/NE1_HR_LC_SR_W_DR/NE1_HR_LC_SR_W_DR.tif")
gg_color_hue <- function(n) {
  hues = seq(15, 375, length = n + 1)
  hcl(h = hues, l = 65, c = 100)[1:n]
}
k <- 3
mapmixture(strmapdata$str[[as.character(k)]], strmapdata$pop, pie_size = 0.5, basemap = earth, boundary = boundary, cluster_cols = gg_color_hue(k), arrow_position = "tr")
}
